Frontend fornecedor incompleto

+++ controle-estoque/app/Http/Controllers/FornecedorController.php
@@ -6,6 +6,7 @@ use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 use App\Http\Controllers\Controller;
 use App\Fornecedor;
+use App\Produto;
 use Illuminate\Support\Facades\DB;
 
 class FornecedorController extends Controller
@@ -23,7 +24,7 @@ class FornecedorController extends Controller
         $fornecedor->endereco = $request->endereco;
         $fornecedor->cnpj = $request->cnpj;
         $fornecedor->save();
-        return redirect('/');
+        return redirect('/fornecedor');
     }
 
     public function editar(Request $request){
@@ -33,19 +34,26 @@ class FornecedorController extends Controller
         $fornecedor->cnpj = $request->cnpj;
         $fornecedor->endereco = $request->endereco;
         $fornecedor->save();
-        return redirect('/produtos');
+        return redirect('/fornecedor');
     }
 
     public function buscar(Request $request){
         parent::validaLogin();
-        $fornecedor = DB::select('select * from fornecedores where nome = ?', [$request->nome]);
+        $fornecedor = DB::select('select * from fornecedores where nome like ?', ['%'.$request->nome.'%']);
         return view('fornecedor.pesquisa', compact('fornecedor'));    
     }
 
     public function deletar(Request $request){
         parent::validaLogin();
-        $fornecedor = Fornecedor::find($request->id)->first();
-        $fornecedor->delete();
-        return redirect('/fornecedores');
+        $fornecedor = Fornecedor::where('id', $request->id)->first();
+        $produtos = DB::select('select * from produtos where fornecedor_id = ?', [$request->id]);
+        if(isset($produtos[0]->id)){
+            return redirect('/fornecedor');
+        }else{
+            $fornecedor->delete();
+            return redirect('/fornecedor');
+        }
+        
+        
     }
 }
